\docType{data}
\name{sb.expression}
\alias{sb.expression}
\title{Gene expression data of 14 patients
A demo dataset containing 6597 gene expression profiles on 14 patients, the corresponding
mutation status is provided in \code{\link{sb.mutation}}}
\format{A data.frame with 6597 rows and 14 variables}
\description{
Gene expression data of 14 patients A demo dataset
containing 6597 gene expression profiles on 14 patients,
the corresponding mutation status is provided in
\code{\link{sb.mutation}}
}
\keyword{datasets}

