\docType{data}
\name{sb.mutation}
\alias{sb.mutation}
\title{Mutation status of 14 patients}
\format{A character vector of 14 elements}
\description{
A character vector indicating the mutation status of 14
patients
}
\references{
Xu, Y., Guo, X., Sun, J. and Zhao. Z. Snowball: resampling
combined with distance-based regression to discover
transcriptional consequences of driver mutation,
manuscript.

Guo, X., Xu, Y. and Zhao, Z.. Driver mutation BRAF
regulates cell proliferation and apoptosis via MITF in the
pathogenesis of melanoma, manuscript.
}
\keyword{datasets}

