\docType{data}
\name{sb.expression}
\alias{sb.expression}
\title{Gene expression data of 14 patients}
\format{A data.frame with 6597 rows and 14 variables}
\description{
A demo dataset containing 6597 gene expression profiles on
14 patients, the corresponding mutation status is provided
in \code{\link{sb.mutation}}
}
\references{
Xu, Y., Guo, X., Sun, J. and Zhao. Z. Snowball: resampling
combined with distance-based regression to discover
transcriptional consequences of driver mutation,
manuscript.

Guo, X., Xu, Y. and Zhao, Z.. Driver mutation BRAF
regulates cell proliferation and apoptosis via MITF in the
pathogenesis of melanoma, manuscript.
}
\keyword{datasets}

