\docType{data}
\name{sb.mutation}
\alias{sb.mutation}
\title{Mutation status of 14 patients
A character vector indicating the mutation status of 14 patients}
\format{A character vector of 14 elements}
\description{
Mutation status of 14 patients A character vector
indicating the mutation status of 14 patients
}
\keyword{datasets}

